+++ resources/views/app.blade.php
@@ -3,7 +3,7 @@
 <head>
     <meta charset="utf-8"/>
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0"/>
-    
+
     <link href="{{ asset('/vendor/fastlane/fonts/inter.css') }}" rel="stylesheet"/>
     <link href="{{ asset('/vendor/fastlane/icons/css/line-awesome.min.css') }}" rel="stylesheet"/>
     <link href="{{ mix('app.css', 'vendor/fastlane') }}" rel="stylesheet"/>