+++ src/EntryTypes/Hooks/OnSavingHook.php
@@ -2,34 +2,38 @@
 
 namespace CbtechLtd\Fastlane\EntryTypes\Hooks;
 
-use CbtechLtd\Fastlane\EntryTypes\EntryType;
+use CbtechLtd\Fastlane\Support\Contracts\EntryInstance as EntryInstanceContract;
+use CbtechLtd\Fastlane\Support\Contracts\EntryType;
 use Illuminate\Database\Eloquent\Model;
 
 class OnSavingHook
 {
-    protected EntryType $entryType;
-    protected Model $model;
+    protected EntryInstanceContract $entryInstance;
     protected array $data;
 
-    public function __construct(EntryType $entryType, Model $model, array $data)
+    public function __construct(EntryInstanceContract $entryInstance, array $data)
     {
-        $this->entryType = $entryType;
-        $this->model = $model;
+        $this->entryInstance = $entryInstance;
         $this->data = $data;
+    }
+
+    public function entryInstance(): EntryInstanceContract
+    {
+        return $this->entryInstance;
     }
 
     public function entryType(): EntryType
     {
-        return $this->entryType;
+        return $this->entryInstance->type();
+    }
+
+    public function model(): Model
+    {
+        return $this->entryInstance->model();
     }
 
     public function data(): array
     {
         return $this->data;
-    }
-
-    public function model(): Model
-    {
-        return $this->model;
     }
 }