+++ src/Support/ControlPanelResources/EntryResource.php
@@ -2,30 +2,34 @@
 
 namespace CbtechLtd\Fastlane\Support\ControlPanelResources;
 
-use CbtechLtd\Fastlane\Support\Contracts\EntryType;
-use CbtechLtd\Fastlane\Support\Schema\Fields\Contracts\WithValue;
+use CbtechLtd\Fastlane\Support\Contracts\EntryInstance;
+use CbtechLtd\Fastlane\Support\Schema\Fields\Contracts\ReadValue;
 use CbtechLtd\JsonApiTransformer\ApiResources\ApiResource;
 use CbtechLtd\JsonApiTransformer\ApiResources\ResourceLink;
 use CbtechLtd\JsonApiTransformer\ApiResources\ResourceMeta;
 use CbtechLtd\JsonApiTransformer\ApiResources\ResourceType;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
 
 class EntryResource extends ResourceType
 {
-    protected EntryType $entryType;
+    protected EntryInstance $entryInstance;
     protected string $destination;
 
-    public function __construct(Model $model, EntryType $entryType)
+    public function __construct(EntryInstance $entryInstance)
     {
-        parent::__construct($model);
-        $this->entryType = $entryType;
+        parent::__construct($entryInstance->model());
+        $this->entryInstance = $entryInstance;
     }
 
     public function type(): string
     {
         return 'fastlane-entry';
+    }
+
+    public function id(): string
+    {
+        return $this->getModel()->getRouteKey() ?? '';
     }
 
     public function transform(): ApiResource
@@ -56,45 +60,51 @@ class EntryResource extends ResourceType
         $fields = $this->getSchemaFields();
 
         return Collection::make($fields)
+            ->filter(fn($field) => $field instanceof ReadValue)
             ->mapWithKeys(
-                fn(WithValue $field) => $field->resolveValue($this->model)
+                fn(ReadValue $field) => $field->readValue($this->entryInstance)->toArray()
             )->all();
     }
 
     protected function meta(): array
     {
         return [
-            ResourceMeta::make('item_label', $this->entryType->makeModelTitle($this->model)),
+            ResourceMeta::make('item_label', $this->entryInstance->title()),
             ResourceMeta::make('entry_type', [
                 'schema'        => Collection::make($this->getSchemaFields()),
                 'panels'        => Collection::make($this->getSchemaPanels()),
-                'singular_name' => $this->entryType->name(),
-                'plural_name'   => $this->entryType->pluralName(),
-                'identifier'    => $this->entryType->identifier(),
-                'icon'          => $this->entryType->icon(),
+                'singular_name' => $this->entryInstance->type()->name(),
+                'plural_name'   => $this->entryInstance->type()->pluralName(),
+                'identifier'    => $this->entryInstance->type()->identifier(),
+                'icon'          => $this->entryInstance->type()->icon(),
             ]),
         ];
     }
 
     protected function links(): array
     {
-        return [
-            ResourceLink::make('self', ["cp.{$this->entryType->identifier()}.edit", $this->model]),
-            ResourceLink::make('parent', ["cp.{$this->entryType->identifier()}.index"]),
+        $links = [
+            ResourceLink::make('parent', ["cp.{$this->entryInstance->type()->identifier()}.index"]),
         ];
+
+        if ($this->entryInstance->model()->exists) {
+            $links[] = ResourceLink::make('self', ["cp.{$this->entryInstance->type()->identifier()}.edit", $this->model]);
+        }
+
+        return $links;
     }
 
     private function getSchemaFields()
     {
         if ($this->destination === 'update') {
-            return $this->entryType->schema()->getUpdateFields();
+            return $this->entryInstance->schema()->getUpdateFields();
         }
 
         if ($this->destination === 'create') {
-            return $this->entryType->schema()->getCreateFields();
+            return $this->entryInstance->schema()->getCreateFields();
         }
 
-        return $this->entryType->schema()->getIndexFields();
+        return $this->entryInstance->schema()->getIndexFields();
     }
 
     private function getSchemaPanels()
@@ -103,6 +113,6 @@ class EntryResource extends ResourceType
             return [];
         }
 
-        return $this->entryType->schema()->getPanels();
+        return $this->entryInstance->schema()->getPanels();
     }
 }