+++ src/Support/Eloquent/BaseModel.php
@@ -10,9 +10,11 @@ use Altek\Eventually\Eventually;
 use CbtechLtd\Fastlane\Support\Eloquent\Concerns\Activable;
 use CbtechLtd\Fastlane\Support\Eloquent\Concerns\Hashable;
 use CbtechLtd\Fastlane\Support\Eloquent\Concerns\RelatesToEntryType;
+use CbtechLtd\Fastlane\Support\Eloquent\Contracts\LoadAttributesFromEntryType;
+use CbtechLtd\Fastlane\Support\Eloquent\Contracts\LoadRelationshipsFromEntryType;
 use Illuminate\Database\Eloquent\Model;
 
-abstract class BaseModel extends Model implements Recordable
+abstract class BaseModel extends Model implements Recordable, LoadAttributesFromEntryType, LoadRelationshipsFromEntryType
 {
     use Hashable, RecordableTrait, Eventually, Activable, RelatesToEntryType;
 }