+++ src/Support/Schema/Fields/Contracts/WithRules.php
@@ -4,13 +4,13 @@ namespace CbtechLtd\Fastlane\Support\Schema\Fields\Contracts;
 
 interface WithRules
 {
-    public function setRules(string $rules): self;
+    public function withRules(string $rules): self;
 
     public function getCreateRules(): array;
 
-    public function setCreateRules(string $rules): self;
+    public function withCreateRules($rules): self;
 
     public function getUpdateRules(): array;
 
-    public function setUpdateRules(string $rules): self;
+    public function withUpdateRules($rules): self;
 }