+++ src/Support/Schema/Fields/FileField.php
@@ -3,7 +3,7 @@
 namespace CbtechLtd\Fastlane\Support\Schema\Fields;
 
 use CbtechLtd\Fastlane\FileAttachment\DraftAttachment;
-use CbtechLtd\Fastlane\Support\Contracts\EntryType as EntryTypeContract;
+use CbtechLtd\Fastlane\Support\Contracts\EntryInstance;
 use CbtechLtd\Fastlane\Support\Schema\Fields\Concerns\ExportsToApiAttribute;
 use CbtechLtd\Fastlane\Support\Schema\Fields\Concerns\HandlesAttachments;
 use CbtechLtd\Fastlane\Support\Schema\Fields\Contracts\ExportsToApiAttribute as ExportsToApiAttributeContract;
@@ -33,14 +33,14 @@ class FileField extends AbstractBaseField implements ExportsToApiAttributeContra
         return $this;
     }
 
-    public function fillModel($model, $value, array $requestData): void
+    public function writeValue(EntryInstance $entryInstance, $value, array $requestData): void
     {
-        if (is_callable($this->fillValueCallback)) {
-            call_user_func($this->fillValueCallback, $model, $value[0], $requestData);
+        if (is_callable($this->writeValueCallback)) {
+            call_user_func($this->writeValueCallback, $entryInstance, $value[0], $requestData);
             return;
         }
 
-        $model->{$this->getName()} = $value[0];
+        $entryInstance->model()->{$this->getName()} = $value[0];
 
         // Check whether files are accepted in the rich editor instance,
         // then persist all draft attachment files.
@@ -48,7 +48,7 @@ class FileField extends AbstractBaseField implements ExportsToApiAttributeContra
             DraftAttachment::persistAllDraft(
                 $draftId,
                 $this,
-                $model,
+                $entryInstance->model(),
             );
         }
     }
@@ -56,18 +56,20 @@ class FileField extends AbstractBaseField implements ExportsToApiAttributeContra
     protected function getTypeRules(): array
     {
         return [
-            $this->getName() => 'file',
+            // TODO: Improve the URL stuff. Something like ImageField and ImageUploader would be better.
+            $this->getName()        => 'array',
+            $this->getName() . '.*' => "sometimes|url|starts_with:" . url('storage/attachments'),
         ];
     }
 
-    protected function resolveConfig(EntryTypeContract $entryType, array $data): array
+    protected function resolveConfig(EntryInstance $entryInstance, string $destination): void
     {
-        return [
+        $this->resolvedConfig = $this->resolvedConfig->merge([
             'fileTypes' => $this->getAccept(),
             'csrfToken' => csrf_token(),
             'links'     => [
-                'self' => URL::relative("cp.{$entryType->identifier()}.attachments", $this->getName()),
+                'self' => URL::relative("cp.{$entryInstance->type()->identifier()}.attachments", $this->getName()),
             ],
-        ];
+        ]);
     }
 }